Simplify block until SIGTERM
Instead of having to write:

```
ic := make(chan os.Signal, 1)
signal.Notify(ic, os.Interrupt, syscall.SIGTERM)
<-ic
c.Close()
```

there is now a process package which contains a block method for
blocking a process until a sigterm arrives to then close a given
io.Closer instead.

This simplifies a infite wait to:

```
process.Block(c)
```

and normalizes the abort conditions.
